refactor(mokumo-shop): relocate CLI direct-path helpers from services/api [#512]
Lifts cli_backup, cli_restore, cli_reset_db, cli_reset_password,
cli_migrate_status and their types (ResetReport, ResetError,
MigrationRecord, MigrateStatusReport, DB_SIDECAR_SUFFIXES) from
services/api/src/lib.rs to crates/mokumo-shop/src/cli.rs.

These are direct-filesystem/direct-DB helpers used by mokumo-server
subcommands as a fallback when the daemon isn't running. They're
NOT duplicates of kikan-cli — kikan-cli's variants go via the admin
UDS (requires a running daemon). Together they implement the "works
with or without a running server" pattern.

mokumo-server repointed to mokumo_shop::cli::*. services/api retains
`pub use mokumo_shop::cli::*` re-exports so BDD tests continue to
compile until B3-bulk migrates them.

Refs #512 (Stage 8, part B2d).
