+++ services/api/src/lib.rs
@@ -23,7 +23,7 @@ pub use mokumo_shop::settings;
 pub use mokumo_shop::setup;
 pub mod ws;
 
-use std::path::{Path, PathBuf};
+use std::path::PathBuf;
 use std::sync::Arc;
 use std::sync::atomic::AtomicBool;
 
@@ -473,331 +473,19 @@ fn build_app_inner(
     (app, ws_handle, state)
 }
 
-/// Reset a user's password directly via SQLite (no server required).
-///
-/// This is the CLI support fallback — opens the database file directly,
-/// hashes the new password with Argon2id, and updates the row.
-/// Returns an error message on failure.
-pub fn cli_reset_password(db_path: &Path, email: &str, new_password: &str) -> Result<(), String> {
-    let conn = rusqlite::Connection::open(db_path)
-        .map_err(|e| format!("Cannot open database at {}: {e}", db_path.display()))?;
-
-    let hash = password_auth::generate_hash(new_password);
-
-    let rows = conn
-        .execute(
-            "UPDATE users SET password_hash = ?1 WHERE email = ?2 AND deleted_at IS NULL",
-            rusqlite::params![hash, email],
-        )
-        .map_err(|e| format!("Failed to update password: {e}"))?;
-
-    if rows == 0 {
-        return Err(format!("No active user found with email '{email}'"));
-    }
-
-    Ok(())
-}
-
 // Binary-shell helpers (process lock, recovery dir, drain timeout) moved to
-// `mokumo_shop::startup` in PR 4b (#512). Re-exports preserved for BDD tests
-// until B3-bulk migrates them.
+// `mokumo_shop::startup` in PR 4b (#512). CLI direct-path helpers
+// (cli_backup, cli_restore, cli_reset_db, cli_reset_password,
+// cli_migrate_status, plus their types) moved to `mokumo_shop::cli`.
+// Re-exports preserved for BDD tests until B3-bulk migrates them.
+pub use mokumo_shop::cli::{
+    DB_SIDECAR_SUFFIXES, MigrateStatusReport, MigrationRecord, ResetError, ResetReport, cli_backup,
+    cli_migrate_status, cli_reset_db, cli_reset_password, cli_restore,
+};
 pub use mokumo_shop::startup::{
     DRAIN_TIMEOUT_SECS, format_lock_conflict_message, format_reset_db_conflict_message,
     lock_file_path, read_lock_info, resolve_recovery_dir, write_lock_info,
 };
-
-/// SQLite sidecar suffixes deleted alongside the main database file.
-pub const DB_SIDECAR_SUFFIXES: &[&str] = &["", "-wal", "-shm", "-journal"];
-
-/// Report from a database reset operation.
-#[derive(Debug, Default)]
-pub struct ResetReport {
-    pub deleted: Vec<PathBuf>,
-    pub not_found: Vec<PathBuf>,
-    pub failed: Vec<(PathBuf, std::io::Error)>,
-    pub recovery_dir_error: Option<(PathBuf, std::io::Error)>,
-    /// Non-fatal: backup directory could not be scanned (only set when `include_backups` is true).
-    pub backup_dir_error: Option<(PathBuf, std::io::Error)>,
-}
-
-/// Fatal errors during database reset (not partial file failures).
-#[derive(Debug, thiserror::Error)]
-pub enum ResetError {
-    #[error(transparent)]
-    Io(#[from] std::io::Error),
-}
-
-/// Delete database files, sidecars, and optionally backups + recovery files.
-///
-/// `profile_dir` is the directory containing `mokumo.db` for the target profile
-/// (e.g. `data_dir/demo` or `data_dir/production`). The caller resolves this
-/// from the `--production` flag before calling.
-///
-/// If `profile_dir` does not exist, all database and backup entries will appear
-/// in `report.not_found`; the function does not return `Err` in this case.
-///
-/// This is a pure filesystem function with no stdin/stdout interaction.
-/// The caller (main.rs) handles confirmation prompts and result display.
-pub fn cli_reset_db(
-    profile_dir: &Path,
-    recovery_dir: &Path,
-    include_backups: bool,
-) -> Result<ResetReport, ResetError> {
-    let mut report = ResetReport::default();
-
-    // 1. Database file + sidecars
-    for suffix in DB_SIDECAR_SUFFIXES {
-        let path = profile_dir.join(format!("mokumo.db{suffix}"));
-        delete_file(&path, &mut report);
-    }
-
-    // 2. Backup files (opt-in)
-    if include_backups {
-        match std::fs::read_dir(profile_dir) {
-            Ok(entries) => {
-                for entry in entries.flatten() {
-                    let name = entry.file_name();
-                    if let Some(name_str) = name.to_str()
-                        && name_str.starts_with("mokumo.db.backup-v")
-                    {
-                        delete_file(&entry.path(), &mut report);
-                    }
-                }
-            }
-            Err(e) if e.kind() == std::io::ErrorKind::NotFound => {
-                // profile_dir doesn't exist — nothing to scan
-            }
-            Err(e) => {
-                report.backup_dir_error = Some((profile_dir.to_path_buf(), e));
-            }
-        }
-    }
-
-    // 3. Recovery directory contents (only mokumo-recovery-*.html files)
-    match std::fs::read_dir(recovery_dir) {
-        Ok(entries) => {
-            for entry in entries.flatten() {
-                let name = entry.file_name();
-                if let Some(name_str) = name.to_str()
-                    && name_str.starts_with("mokumo-recovery-")
-                    && name_str.ends_with(".html")
-                {
-                    delete_file(&entry.path(), &mut report);
-                }
-            }
-        }
-        Err(e) if e.kind() == std::io::ErrorKind::NotFound => {}
-        Err(e) => {
-            report.recovery_dir_error = Some((recovery_dir.to_path_buf(), e));
-        }
-    }
-
-    Ok(report)
-}
-
-/// Create a manual backup of the database using the SQLite Online Backup API.
-///
-/// Resolves the output path: if `output` is provided, uses it directly; otherwise
-/// generates a timestamped filename in the database's directory.
-///
-/// This is safe to run while the server is running — the Online Backup API
-/// handles WAL mode and concurrent access correctly.
-pub fn cli_backup(
-    db_path: &Path,
-    output: Option<&Path>,
-) -> Result<kikan::backup::BackupResult, String> {
-    let output_path = match output {
-        Some(p) => p.to_path_buf(),
-        None => {
-            let dir = db_path.parent().unwrap_or(Path::new("."));
-            dir.join(kikan::backup::build_timestamped_name())
-        }
-    };
-
-    let result = kikan::backup::create_backup(db_path, &output_path).map_err(|e| format!("{e}"))?;
-
-    kikan::backup::verify_integrity(&output_path)
-        .map_err(|e| format!("Backup created but integrity check failed: {e}"))?;
-
-    // Bundle the shop logo as a sibling file alongside the backup DB.
-    // Read from the backup file (output_path) to match the state we just captured.
-    // Failure is non-fatal — log a warning and continue.
-    let production_dir = db_path.parent().unwrap_or(Path::new("."));
-    if let Ok(conn) = rusqlite::Connection::open(&output_path)
-        && let Ok(ext) = conn.query_row(
-            "SELECT logo_extension FROM shop_settings WHERE id = 1 AND logo_extension IS NOT NULL",
-            [],
-            |row| row.get::<_, String>(0),
-        )
-    {
-        let logo_src = production_dir.join(format!("logo.{ext}"));
-        let logo_dst = output_path.with_extension(format!("logo.{ext}"));
-        if let Err(e) = std::fs::copy(&logo_src, &logo_dst) {
-            tracing::warn!(
-                "cli_backup: could not copy logo file {:?} → {:?}: {e}",
-                logo_src,
-                logo_dst
-            );
-        }
-    }
-
-    Ok(result)
-}
-
-/// Restore the database from a backup file.
-///
-/// Verifies the backup's integrity, creates a safety backup of the current
-/// database, then overwrites it with the backup contents.
-///
-/// The caller must hold the process lock (server must not be running).
-pub fn cli_restore(
-    db_path: &Path,
-    backup_path: &Path,
-) -> Result<kikan::backup::RestoreResult, String> {
-    let result = kikan::backup::restore_from_backup(db_path, backup_path, DB_SIDECAR_SUFFIXES)
-        .map_err(|e| format!("{e}"))?;
-
-    // Restore the shop logo from its sibling file, if present.
-    // First sweep any stale logo.* files so a changed extension doesn't leave orphans.
-    // Failure is non-fatal — log a warning and continue.
-    let production_dir = db_path.parent().unwrap_or(Path::new("."));
-    for candidate_ext in &["png", "jpeg", "webp"] {
-        let stale = production_dir.join(format!("logo.{candidate_ext}"));
-        if stale.exists()
-            && let Err(e) = std::fs::remove_file(&stale)
-        {
-            tracing::warn!("cli_restore: could not remove stale logo {:?}: {e}", stale);
-        }
-    }
-    if let Ok(conn) = rusqlite::Connection::open(backup_path)
-        && let Ok(ext) = conn.query_row(
-            "SELECT logo_extension FROM shop_settings WHERE id = 1 AND logo_extension IS NOT NULL",
-            [],
-            |row| row.get::<_, String>(0),
-        )
-    {
-        let sibling = backup_path.with_extension(format!("logo.{ext}"));
-        if sibling.exists() {
-            let logo_dst = production_dir.join(format!("logo.{ext}"));
-            if let Err(e) = std::fs::copy(&sibling, &logo_dst) {
-                tracing::warn!(
-                    "cli_restore: could not restore logo file {:?} → {:?}: {e}",
-                    sibling,
-                    logo_dst
-                );
-            }
-        }
-    }
-
-    Ok(result)
-}
-
-/// A single migration record from `seaql_migrations`, with computed status.
-#[derive(Debug)]
-pub struct MigrationRecord {
-    pub name: String,
-    pub applied_at: Option<chrono::DateTime<chrono::Utc>>,
-}
-
-/// Output of `mokumo migrate status`.
-#[derive(Debug)]
-pub struct MigrateStatusReport {
-    pub current_version: Option<String>,
-    pub applied: Vec<MigrationRecord>,
-    pub pending: Vec<String>,
-    /// Migrations recorded in the DB but not known to this binary.
-    /// Non-empty only on binary downgrade — the schema is ahead of the binary.
-    pub unknown: Vec<String>,
-}
-
-/// Query the migration state of a database file.
-///
-/// Opens the database with a raw rusqlite connection (no pool, no migrations).
-/// Returns the set of applied migrations (with timestamps) and pending migrations
-/// (known to the binary but not recorded in `seaql_migrations`).
-///
-/// Returns an error string on any database or query failure.
-pub fn cli_migrate_status(db_path: &Path) -> Result<MigrateStatusReport, String> {
-    let conn = rusqlite::Connection::open_with_flags(
-        db_path,
-        rusqlite::OpenFlags::SQLITE_OPEN_READ_ONLY | rusqlite::OpenFlags::SQLITE_OPEN_URI,
-    )
-    .map_err(|e| format!("Cannot open database at {}: {e}", db_path.display()))?;
-
-    let table_exists: bool = conn
-        .query_row(
-            "SELECT COUNT(*) > 0 FROM sqlite_master WHERE type='table' AND name='seaql_migrations'",
-            [],
-            |row| row.get(0),
-        )
-        .map_err(|e| format!("Failed to query sqlite_master: {e}"))?;
-
-    if !table_exists {
-        let known = mokumo_shop::db::known_migration_names();
-        return Ok(MigrateStatusReport {
-            current_version: None,
-            applied: vec![],
-            pending: known,
-            unknown: vec![],
-        });
-    }
-
-    let mut stmt = conn
-        .prepare("SELECT version, applied_at FROM seaql_migrations ORDER BY version")
-        .map_err(|e| format!("Failed to prepare migration query: {e}"))?;
-
-    let applied: Vec<MigrationRecord> = stmt
-        .query_map([], |row| {
-            let name: String = row.get(0)?;
-            let ts: i64 = row.get(1)?;
-            Ok((name, ts))
-        })
-        .map_err(|e| format!("Failed to query seaql_migrations: {e}"))?
-        .map(|r| {
-            r.map(|(name, ts)| MigrationRecord {
-                applied_at: chrono::DateTime::from_timestamp(ts, 0),
-                name,
-            })
-        })
-        .collect::<Result<_, _>>()
-        .map_err(|e: rusqlite::Error| format!("Failed to read migration row: {e}"))?;
-
-    let known = mokumo_shop::db::known_migration_names();
-    let known_set: std::collections::HashSet<&str> = known.iter().map(|n| n.as_str()).collect();
-
-    let unknown: Vec<String> = applied
-        .iter()
-        .filter(|r| !known_set.contains(r.name.as_str()))
-        .map(|r| r.name.clone())
-        .collect();
-
-    let applied_names: std::collections::HashSet<&str> =
-        applied.iter().map(|r| r.name.as_str()).collect();
-
-    let pending: Vec<String> = known
-        .into_iter()
-        .filter(|n| !applied_names.contains(n.as_str()))
-        .collect();
-
-    let current_version = applied.last().map(|r| r.name.clone());
-
-    Ok(MigrateStatusReport {
-        current_version,
-        applied,
-        pending,
-        unknown,
-    })
-}
-
-fn delete_file(path: &Path, report: &mut ResetReport) {
-    match std::fs::remove_file(path) {
-        Ok(()) => report.deleted.push(path.to_path_buf()),
-        Err(e) if e.kind() == std::io::ErrorKind::NotFound => {
-            report.not_found.push(path.to_path_buf());
-        }
-        Err(e) => report.failed.push((path.to_path_buf(), e)),
-    }
-}
 
 #[cfg(debug_assertions)]
 async fn debug_recovery_dir(State(state): State<SharedState>) -> impl IntoResponse {
